collector-service/requirements.txt
fastapi
uvicorn
pydantic
boto3
python-dotenv
paho-mqtt